nsible-boilerplate: Try to get the "upgrade" logic right
Don't try to install the Python package "ansible" and the Ansible
collection "alexbarton.boilerplate" first, depend on the Python
requirements.txt and Ansible Galaxy requirements.yml files for this.

This enables the user to specify the sources (and versions) that are
required for the specific project (and no longer requires the environment
variable BOILERPLATE_COLLECTION_SRC to be set, for example).