ansible-boilerplate: Fix changing the working directory
The idea is to change to the project directory, which is good.

But the current logic, introduced in commit e61c67d, is broken(!) and
only works when the script is called from the "bin" directory of an
already initialized project.

It fails horribly when called from the "galaxy" directory, which is
required to initialize a new project in the first place! Ooooops!

This patch hopefully works for both locations and "all" cases ...